finished sign up, login, and App.js
