+++ package.json
@@ -5,22 +5,22 @@
   "dependencies": {
     "downloadbutton": "^1.0.0",
     "firebase": "^4.5.0",
-    "firebaseui": "^2.3.0",
+    "firebaseui": "^2.4.0",
     "geofire": "^4.1.2",
     "mapbox-gl": "^0.40.1",
-    "material-ui": "^0.19.3",
+    "material-ui": "^0.19.4",
     "prop-types": "^15.6.0",
     "react": "^16.0.0",
     "react-bootstrap": "^0.31.3",
-    "react-burger-menu": "^2.1.6",
+    "react-burger-menu": "^2.1.8",
     "react-dom": "^16.0.0",
     "react-file-download": "^0.3.5",
     "react-geolocation": "^1.0.3",
     "react-google-button": "^0.3.0",
-    "react-google-maps": "^8.3.4",
-    "react-icons": "^2.2.5",
+    "react-google-maps": "^8.3.5",
+    "react-icons": "^2.2.7",
     "react-map-gl": "^3.0.5",
-    "react-mapbox-gl": "^2.5.2",
+    "react-mapbox-gl": "^2.7.0",
     "react-notification-system": "^0.2.15",
     "react-redux": "^5.0.6",
     "react-redux-firebase": "^1.5.0",
