+++ src/components/LoginSignup/style.css
@@ -1,0 +1,8 @@
+.signupComp {
+    margin: 4px 4px 4px 4px;
+}
+
+.loginComp {
+    margin: 5px 5px 5px 5px;
+
+}
