+++ src/index.css
@@ -78,3 +78,7 @@ a {
 }
 
 
+.mainComp {
+  backgroundColor: black;
+  maring-left: 25%;
+}
