+++ src/index.js
@@ -6,10 +6,13 @@ import './index.css'
 import Main from './components/Main'
 import { BrowserRouter as Router } from 'react-router-dom'
 import registerServiceWorker from './registerServiceWorker'
+import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider'
 
 ReactDOM.render(
 	<Provider store={store}>
-		<Main />
+		<MuiThemeProvider>
+			<Main />
+		</MuiThemeProvider>
 	</Provider>,
 	document.getElementById('root'))
 registerServiceWorker()
